Fix improper link, and move a section to the end
The "Changes on 'client_id' and 'resource' arguments after 0.1.0" section was added 1 year ago as an explanation of a change after version 0.1.0. Now, it feels somewhat weird to have this section show up early in this README. We decide to demote this paragraph to the very end of README, in favor of drawing user's attention to the "Samples and Documentation" section instead.