Add an enable_pii switch
We do implement an enable_pii switch in this commit, and it will control
the behaviors implemented by subsequent commits in this PR.

The enable_pii switch is not implemented as a publicly available switch,
though. Otherwise the implication would become "all direct and indirect
ADAL Python callers would need to chain up and relay such switch to ADAL
Python". That won't be easy to coordinate.

Instead, we expect this switch would remain the default False,
so no PII would show up in the logs at all. In the extremely rare
customer-support case that we would like to know the user account for
debug, we typically get it from the support case description anyway.